Refactor tests to use new `ReplicatorTestsBase` class.
Moved replicator logic from `NotificationTestsBase` to a new `ReplicatorTestsBase` class to enhance modularity. Updated test files and renamed them to reflect their dependency on the new base class.
