+++ test/LionWeb.Core.Test/Notification/Replicator/NotificationTests_Twoway.cs
@@ -22,7 +22,7 @@ using Core.Notification.Pipe;
 using Languages.Generated.V2024_1.Shapes.M2;
 
 [TestClass]
-public class NotificationTests_Twoway: NotificationTestsBase 
+public class NotificationTests_Twoway: ReplicatorTestsBase 
 {
     #region Properties
